// src/app-types.ts
export interface Barber {
  image?: string;
  id: number;
  name: string;
  company_id: number;
  specialization: string;
  information: string;
  api_id?: any;
  fired: number;
  is_fired: boolean;
  dismissal_date?: any;
  dismissal_reason?: any;
  hidden: number;
  is_online: boolean;
  status: number;
  is_deleted: boolean;
  user_id: number;
  rating: number;
  prepaid: string;
  is_chain: boolean;
  weight: number;
  is_rating_shown: boolean;
  is_online_random_choice_allowed: boolean;
  seance_step: number;
  seance_search_step: number;
  seance_search_start: number;
  seance_search_finish: number;
  avatar: string;
  avatar_big: string;
  user: User;
}

export interface User {
  id: number;
  name: string;
  phone: string;
  avatar: string;
  email: string;
}